libtaskmanager: Improve guarding of invalid proxies
We have some issues with something getting out of sync in the taskmanager grouping model based on bug reports.

There's some slightly off error handling here. It both doesn't make a crash
visible at the right place, and if we do want to ignore invalid proxies, we should do it correctly.

This is a blind fix, I haven't reproduced the issue.

https://crash-reports.kde.org/organizations/kde/issues/221251