BadgeOverlay: Fit the text better and improve contrast
The old appearance was too wide in some circumstances. Now it tries to fit the text, be circular or stay within the bounds of the system tray icon.

The text had poor readability. Now the text is slightly more readable.

The shadow on the dark version was basically invisible. Now it has a separator colored outline.