wallpapers/image: avoid recursive rearrange in SlideshowComponent
The (settings for the) slideshow component has an item in a layout that
sets its preferred width according to the layout's width. This causes a
whole slew of error messages to be printed upon loading, as RowLayout
complains about recursive rearranges.

Instead, we base it on that Layout's parent layout, which is sized
externally and the RowLayout is fillWidth within the parent ColumnLayout
(with no extra horizontal padding) so their size will eventually match
anyway, just with fewer extra steps. This resolves the runtime warnings
when loading the slideshow settings.